+++ R/layout.R
@@ -88,6 +88,8 @@ rangeslider <- function(p, start = NULL, end = NULL, ...) {
 #' @param p a plotly object
 #' @param ... these arguments are documented at 
 #' \url{https://github.com/plotly/plotly.js/blob/master/src/plot_api/plot_config.js}
+#' @param cloud deprecated. Use `showSendToCloud` instead.
+#' @param showSendToCloud include the send data to cloud button?
 #' @param locale locale to use. See [here](https://github.com/plotly/plotly.js/tree/master/dist#to-include-localization) for more info.
 #' @param mathjax add [MathJax rendering support](https://github.com/plotly/plotly.js/tree/master/dist#to-support-mathjax).
 #' If `"cdn"`, mathjax is loaded externally (meaning an internet connection is needed for 
@@ -129,7 +131,7 @@ rangeslider <- function(p, start = NULL, end = NULL, ...) {
 #' config(p, locale = "zh-CN")
 #' 
 
-config <- function(p, ..., locale = NULL, mathjax = NULL) {
+config <- function(p, ..., cloud = FALSE, showSendToCloud = cloud, locale = NULL, mathjax = NULL) {
   
   if (!is.null(locale)) {
     p$dependencies <- c(
@@ -157,6 +159,8 @@ config <- function(p, ..., locale = NULL, mathjax = NULL) {
   args <- list(...)
   if ("collaborate" %in% names(args)) warning("The collaborate button is no longer supported")
   p$x$config <- modify_list(p$x$config, args)
+  if (cloud) message("The `cloud` argument is deprecated. Use `showSendToCloud` instead.")
+  p$x$config$showSendToCloud <- showSendToCloud
 
   p
 }